exercise 8 HITL Contract Redlining Orchestrator-Tested/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.20
python-docx==1.2.0
reportlab==4.4.4
